Relocated vectors to its own module
