+++ zmats/zmats.py
@@ -32,6 +32,10 @@ import re
 from rmgpy.molecule.molecule import Atom, Molecule
 
 from zmats.exceptions import ZMatError
+from zmats.vectors import (calculate_angle,
+                           calculate_dihedral_angle,
+                           calculate_distance,
+                           get_vector_length)
 
 
 DEFAULT_CONSOLIDATION_R_TOL = 1e-4
@@ -1658,7 +1662,7 @@ def compare_zmats(zmat1: dict,
     """
     Compare two zmats. The zmats must have identical variables (i.e., derived from the same connectivity or ordered xyz,
     using the same constraints).
-    This function does not make use of the zmat map, but does check that it is the identical.
+    This function does not make use of the zmat map, but does check that it is identical.
 
     Args:
         zmat1 (dict): zmat1.
@@ -1809,216 +1813,3 @@ def determine_top_group_indices(mol: Molecule,
             explored_atom_list.append(atom3)  # mark as explored
         atom_list_to_explore1, atom_list_to_explore2 = atom_list_to_explore2, []
     return top
-
-
-# Vectors:
-
-def calculate_distance(coords: list or tuple,
-                       atoms: list,
-                       ) -> float:
-    """
-    Calculate a distance.
-
-    Args:
-        coords (list, tuple): The array-format or tuple-format coordinates.
-        atoms (list): The 2 atoms to calculate defining the vector for which the length will be calculated (0-indexed).
-
-    Returns:
-        float: The distance in the coords units.
-
-    Raises:
-        ZMatError: If ``index`` is out of range, or ``atoms`` is of wrong length or has repeating indices.
-        TypeError: If ``coords`` is of wrong type.
-    """
-    if isinstance(coords, dict) and 'coords' in coords:
-        coords = coords['coords']
-    if not isinstance(coords, (list, tuple)):
-        raise TypeError(f'coords must be a list or a tuple, got\n{coords}\nwhich is a {type(coords)}')
-    if len(atoms) != 2:
-        raise ZMatError(f'distance atom list must be of length two, got {len(atoms)}')
-    if len(set(atoms)) < 2:
-        raise ZMatError(f'some atoms are repetitive: {atoms}')
-    new_atoms = list()
-    for atom in atoms:
-        if isinstance(atom, str) and 'X' in atom:
-            new_atoms.append(int(atom[1:]))
-        else:
-            new_atoms.append(atom)
-    if not all([isinstance(a, int) for a in new_atoms]):
-        raise ZMatError(f'all entries in atoms must be integers, got: {new_atoms} ({[type(a) for a in new_atoms]})')
-    coords = np.asarray(coords, dtype=np.float32)
-    vector = coords[new_atoms[1]] - coords[new_atoms[0]]
-    return get_vector_length(vector)
-
-
-def calculate_angle(coords: list or tuple,
-                    atoms: list,
-                    units: str = 'degs',
-                    ) -> float:
-    """
-    Calculate an angle.
-
-    Args:
-        coords (list, tuple): The array-format or tuple-format coordinates.
-        atoms (list): The 3 atoms defining the angle (0-indexed).
-        units (str, optional): The desired units, either 'rads' for radians, or 'degs' for degrees.
-
-    Returns:
-        float: The angle.
-
-    Raises:
-        ZMatError: If ``index`` is out of range, or ``atoms`` is of wrong length or has repeating indices.
-        TypeError: If ``coords`` is of wrong type.
-    """
-    if isinstance(coords, dict) and 'coords' in coords:
-        coords = coords['coords']
-    if not isinstance(coords, (list, tuple)):
-        raise TypeError(f'coords must be a list or a tuple, got\n{coords}\nwhich is a {type(coords)}')
-    if len(atoms) != 3:
-        raise ZMatError(f'angle atom list must be of length three, got {len(atoms)}')
-    if len(set(atoms)) < 3:
-        raise ZMatError(f'some atoms are repetitive: {atoms}')
-    new_atoms = list()
-    for atom in atoms:
-        if isinstance(atom, str) and 'X' in atom:
-            new_atoms.append(int(atom[1:]))
-        else:
-            new_atoms.append(atom)
-    if not all([isinstance(a, int) for a in new_atoms]):
-        raise ZMatError(f'all entries in atoms must be integers, got: {new_atoms} ({[type(a) for a in new_atoms]})')
-    coords = np.asarray(coords, dtype=np.float32)
-    v1 = coords[new_atoms[1]] - coords[new_atoms[0]]
-    v2 = coords[new_atoms[1]] - coords[new_atoms[2]]
-    return get_angle(v1, v2, units=units)
-
-
-def calculate_dihedral_angle(coords: list or tuple,
-                             torsion: list,
-                             units: str = 'degs'):
-    """
-    Calculate a dihedral angle.
-
-    Args:
-        coords (list, tuple): The array-format or tuple-format coordinates.
-        torsion (list): The 4 atoms defining the dihedral angle (0-indexed).
-        units (str, optional): The desired units, either 'rads' for radians, or 'degs' for degrees.
-
-    Returns:
-        float: The dihedral angle in a 0-360 degrees range.
-
-    Raises:
-        ZMatError: If ``index`` is out of range, or ``torsion`` is of wrong length or has repeating indices.
-        TypeError: If ``coords`` is of wrong type.
-    """
-    if isinstance(coords, dict) and 'coords' in coords:
-        coords = coords['coords']
-    if not isinstance(coords, (list, tuple)):
-        raise TypeError(f'coords must be a list or a tuple, got\n{coords}\nwhich is a {type(coords)}')
-    if len(torsion) != 4:
-        raise ZMatError(f'torsion atom list must be of length four, got {len(torsion)}')
-    if len(set(torsion)) < 4:
-        raise ZMatError(f'some indices in torsion are repetitive: {torsion}')
-    new_torsion = list()
-    for atom in torsion:
-        if isinstance(atom, str) and 'X' in atom:
-            new_torsion.append(int(atom[1:]))
-        else:
-            new_torsion.append(atom)
-    if not all([isinstance(t, int) for t in new_torsion]):
-        raise ZMatError(f'all entries in torsion must be integers, got: {new_torsion} '
-                        f'({[type(t) for t in new_torsion]})')
-    coords = np.asarray(coords, dtype=np.float32)
-    v1 = coords[new_torsion[1]] - coords[new_torsion[0]]
-    v2 = coords[new_torsion[2]] - coords[new_torsion[1]]
-    v3 = coords[new_torsion[3]] - coords[new_torsion[2]]
-    return get_dihedral(v1, v2, v3, units=units)
-
-
-def unit_vector(vector: list) -> list:
-    """
-    Calculate a unit vector in the same direction as the input vector.
-
-    Args:
-        vector (list): The input vector.
-
-    Returns:
-        list: The unit vector.
-    """
-    length = get_vector_length(vector)
-    return [vi / length for vi in vector]
-
-
-def get_vector_length(v: list) -> float:
-    """
-    Get the length of an ND vector
-
-    Args:
-        v (list): The vector.
-
-    Returns:
-        float: The vector's length.
-    """
-    return float(np.dot(v, v) ** 0.5)
-
-
-def get_angle(v1: list,
-              v2: list,
-              units: str = 'degs',
-              ) -> float:
-    """
-    Calculate the angle between two vectors.
-
-    Args:
-         v1 (list): Vector 1.
-         v2 (list): Vector 2.
-         units (str, optional): The desired units, either 'rads' for radians, or 'degs' for degrees.
-
-    Returns:
-        float: The angle between ``v1`` and ``v2`` in the desired units.
-
-    Raises:
-        ZMatError: If ``v1`` and ``v2`` are of different lengths.
-    """
-    if len(v1) != len(v2):
-        raise ZMatError(f'v1 and v2 must be the same length, got {len(v1)} and {len(v2)}.')
-    v1_u, v2_u = unit_vector(v1), unit_vector(v2)
-    conversion = 180 / math.pi if 'degs' in units else 1
-    return float(np.arccos(np.clip(np.dot(v1_u, v2_u), -1.0, 1.0)) * conversion)
-
-
-def get_dihedral(v1: list,
-                 v2: list,
-                 v3: list,
-                 units: str = 'degs',
-                 ) -> float:
-    """
-    Calculate the dihedral angle between three vectors.
-    ``v2`` connects between ``v1`` and ``v3``.
-    Inspired by ASE Atoms.get_dihedral().
-
-    Args:
-         v1 (list): Vector 1.
-         v2 (list): Vector 2.
-         v3 (list): Vector 3.
-         units (str, optional): The desired units, either 'rads' for radians, or 'degs' for degrees.
-
-    Returns:
-        float: The dihedral angle between ``v1`` and ``v2`` in the desired units.
-
-    Raises:
-        ZMatError: If either ``v1`` or ``v2`` have lengths different than three.
-    """
-    if len(v1) != 3 or len(v2) != 3 or len(v3) != 3:
-        raise ZMatError(f'v1, v2, and v3 must have a length of three, got {len(v1)}, {len(v2)}, and {len(v3)}.')
-    v1, v2, v3 = np.array(v1, np.float64), np.array(v2, np.float64), np.array(v3, np.float64)
-    v2_x_v1 = np.cross(v2, v1)
-    v2_x_v1 /= float(np.linalg.norm(v2_x_v1))
-    v3_x_v2 = np.cross(v3, v2)
-    v3_x_v2 /= np.linalg.norm(v3_x_v2)
-    dihedral = np.arccos(np.clip(np.vdot(v2_x_v1, v3_x_v2), -1, 1))
-    if np.vdot(v2_x_v1, v3) > 0:
-        dihedral = 2 * np.pi - dihedral
-    if np.isnan(dihedral):
-        raise ZMatError('Could not calculate a dihedral angle')
-    conversion = 180 / math.pi if 'degs' in units else 1
-    return float(dihedral * conversion)
